refactor(app): relocate TouchControlButton
